[apply suggestion] https://review.udacity.com/#!/reviews/542369
When you run the autonomous mode you can see on the left side there is
a view from the camera and on top of it there is a black rectangle
showing nothing.

In order to show color thresholded image multiply the right hand of
assignment by 255 to turn 0s and 1s into 0s and 255s as expected by
Rover.vision_image.